feat: Enhance user preferences management in admin panel
- Updated userPreferencesRouter to simplify the endpoint for fetching user preferences by removing the facilityId parameter.
- Introduced useAdminUserPreferencesQuery hook for querying admin-specific user preferences.
- Refactored existing user preferences query logic to align with the new endpoint structure.
- Updated relevant components to utilize the new admin user preferences query for improved data handling.
