+++ packages/central-server/app/admin/userPreferences.js
@@ -3,26 +3,23 @@ import asyncHandler from 'express-async-handler';
 
 export const userPreferencesRouter = express.Router();
 
-// GET /admin/user/userPreferences/:facilityId
 userPreferencesRouter.get(
-  '/userPreferences/:facilityId',
+  '/userPreferences',
   asyncHandler(async (req, res) => {
     const {
       store: {
         models: { UserPreference },
       },
       user: currentUser,
-      params: { facilityId },
     } = req;
 
     req.checkPermission('read', currentUser);
 
-    const userPreferences = await UserPreference.getAllPreferences(currentUser.id, facilityId);
+    const userPreferences = await UserPreference.getAllPreferences(currentUser.id, '');
     res.send(userPreferences || {});
   }),
 );
 
-// POST /admin/user/userPreferences
 userPreferencesRouter.post(
   '/userPreferences',
   asyncHandler(async (req, res) => {
