+++ packages/web/app/api/suggesters.js
@@ -13,10 +13,7 @@ export const useSuggester = (type, options) => {
     () =>
       new Suggester(api, type, {
         ...options,
-        baseQueryParameters: { 
-          ...(facilityId && facilityId.trim() && { facilityId }), 
-          ...options?.baseQueryParameters 
-        },
+        baseQueryParameters: { facilityId, ...options?.baseQueryParameters },
       }),
     [api, type, facilityId, options],
   );
